refactor: 디렉토리 구조 변경 및 사용처 적용
