+++ src/components/common/Badges/CounterBadge.tsx
